Refactor to isolate Vapor specific code
