+++ Sources/Orders/OrdersService.swift
@@ -5,12 +5,12 @@
 //  Created by Francesco Paolo Severino on 01/07/24.
 //
 
-import Vapor
+import Foundation
 import FluentKit
 
 /// The main class that handles Wallet orders.
 public final class OrdersService: Sendable {
-    private let service: OrdersServiceCustom<Order, OrdersDevice, OrdersRegistration, OrdersErrorLog>
+    let service: OrdersServiceCustom<Order, OrdersDevice, OrdersRegistration, OrdersErrorLog>
     
     /// Initializes the service.
     ///
@@ -19,15 +19,6 @@ public final class OrdersService: Sendable {
     ///   - logger: The `Logger` to use.
     public init(delegate: any OrdersDelegate, logger: Logger? = nil) throws {
         service = try .init(delegate: delegate, logger: logger)
-    }
-
-    /// Registers all the routes required for Wallet orders to work.
-    ///
-    /// - Parameters:
-    ///   - app: The `Vapor.Application` to setup the routes.
-    ///   - pushMiddleware: The `Middleware` to use for push notification routes. If `nil`, push routes will not be registered.
-    public func registerRoutes(app: Application, pushMiddleware: (any Middleware)? = nil) {
-        service.registerRoutes(app: app, pushMiddleware: pushMiddleware)
     }
 
     /// Generates the order content bundle for a given order.
